requirements.txt
pandas>=1.5
requests>=2.31
PyYAML>=6.0
jsonschema>=4.17
tqdm>=4.66
tenacity>=8.2
pydantic>=2.0
requests-cache>=1.1

# Development
pytest>=7.4
requests-mock>=1.11
black>=23.0
ruff>=0.1
mypy>=1.4
pydantic>=2.0
types-requests>=2.32
types-PyYAML>=6.0
types-jsonschema>=4.25
pandas-stubs>=2.3
